# openapi/sig_agent_connector.yaml
swagger: '2.0'
info:
  title: SIG Compliance Assistant
  description: >-
    Multi-agent engine that answers Standardized Information Gathering (SIG)
    questionnaire items for a SOC2 FinTech vendor, grounded in SOPs and
    historical SIGs. Power Platform custom connectors require Swagger 2.0.
  version: '1.0.0'
host: REPLACE_WITH_YOUR_HOST          # e.g. sig-agent.azurewebsites.net
basePath: /
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  api_key:
    type: apiKey
    in: header
    name: x-api-key
security:
  - api_key: []
paths:
  /health:
    get:
      summary: Health check
      operationId: Health
      responses:
        '200':
          description: Service is up
  /answer:
    post:
      summary: Answer a single SIG question
      operationId: AnswerQuestion
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/AskRequest'
      responses:
        '200':
          description: A grounded, cited answer
          schema:
            $ref: '#/definitions/Answer'
        '401':
          description: Invalid or missing API key
  /answer-batch:
    post:
      summary: Answer multiple SIG questions
      operationId: AnswerBatch
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/BatchRequest'
      responses:
        '200':
          description: A list of grounded answers
          schema:
            type: object
            properties:
              answers:
                type: array
                items:
                  $ref: '#/definitions/Answer'
definitions:
  AskRequest:
    type: object
    required: [question]
    properties:
      question:
        type: string
        description: A single SIG questionnaire item.
  BatchRequest:
    type: object
    required: [questions]
    properties:
      questions:
        type: array
        items:
          type: string
  Citation:
    type: object
    properties:
      document_id: { type: string }
      section: { type: string }
      version: { type: string }
      effective_date: { type: string }
      scope: { type: string }
      url: { type: string }
  Answer:
    type: object
    properties:
      question: { type: string }
      answer: { type: string }
      status:
        type: string
        description: validated_from_history | synthesized_new | escalated
      needs_human_review: { type: boolean }
      confidence: { type: number }
      citations:
        type: array
        items:
          $ref: '#/definitions/Citation'
      notes: { type: string }